add division and decimal input

+++ app/src/main/java/io/github/anodoze/mathwords/QuizState.kt
@@ -3,5 +3,5 @@ package io.github.anodoze.mathwords
 sealed class QuizState {
     object Loading : QuizState()
     data class Awaiting(val card: Card, val input: String) : QuizState()
-    data class Wrong(val card: Card, val correctAnswer: Int) : QuizState()
+    data class Wrong(val card: Card, val correctAnswer: Float) : QuizState()
 }
